+++ lib/eson/language.rb
@@ -1,5 +1,7 @@
 require 'forwardable'
 require 'pry'
+require_relative './lexeme_capture.rb'
+require_relative './ebnf.rb'
 
 module Eson
 
@@ -26,151 +28,6 @@ module Eson
         rule_list = rule_seq.map{|i| i.to_s}
         "#{self.class.to_s.gsub(/Struct::/, "")} has the following production rules:\n#{rule_list.join("\n")}"
       end              
-    end
-
-    #Operations and data structures for the lexeme field
-    #  of Eson::Language::RuleSeq::Rule. Token has a
-    #  regexp that matches a fixed lexeme or a set of strings. 
-    module LexemeCapture
-
-      LexemeTypeError = Class.new(StandardError)
-      
-      Token = Struct.new :lexeme, :name, :alternation_names, :line_number
-
-      def make_token(lexeme)
-        if lexeme.instance_of? Symbol
-          Token.new(lexeme, @name)
-        elsif lexeme.instance_of? String
-          Token.new(lexeme.intern, @name)
-        else
-          raise LexemeTypeError, lexeme_type_error_message(lexeme)
-        end
-      end
-
-      def lexeme_type_error_message(lexeme)
-        "Lexeme provided to method #{caller_locations[0].label} must be either a Symbol or a String but the given lexeme - #{lexeme} is a #{lexeme.class}."
-      end
-
-      def match_token(string)
-        lexeme = self.match(string).to_s.intern
-        self.make_token(lexeme)
-      end
-      
-      def match(string)
-        string.match(rxp)
-      end
-      
-      def rxp
-        apply_at_start(@start_rxp)
-      end
-      
-      def match_rxp?(string)
-        regex_match?(self.rxp, string)
-      end
-
-      def match_start(string)
-        if self.nonterminal?
-          string.match(@start_rxp)
-        else
-          nil
-        end
-      end
-
-      def regex_match?(regex, string)
-        #does not catch zero or more matches that return "", the empty string
-        (string =~ apply_at_start(regex)).nil? ? false : true
-      end      
-
-      def apply_at_start(regex)
-        /\A#{regex.source}/
-      end
-    end
-
-    #Operations and data structures for the ebnf field
-    #  of the Eson::Language::RuleSeq::Rule
-    module EBNF
-
-      Terminal = Struct.new(:rule_name)
-      NonTerminal = Struct.new(:rule_name)
-      
-      ConcatenationRule = Struct.new(:term_list)
-      AlternationRule = Struct.new(:term_set)
-      RepetitionRule = Struct.new(:term)
-      OptionRule = Struct.new(:term)
-
-      def terminal?
-        self.ebnf.nil?
-      end
-      
-      def nonterminal?
-        !terminal?
-      end
-
-      def nullable?
-        if self.option_rule? || self.repetition_rule?
-          true
-        elsif @first_set.include? :nullable
-          true
-        else
-          false
-        end
-      end
-
-      def term_names
-        if self.terminal?
-          nil
-        elsif alternation_rule?
-          self.ebnf.term_set.map{|i| i.rule_name}
-        elsif concatenation_rule?
-          self.ebnf.term_list.map{|i| i.rule_name}
-        elsif repetition_rule? || option_rule?
-          [ebnf.term.rule_name]
-        end
-      end
-
-      #FIXME this no longer works as terminals which have been
-      #converted from nonterminals have an undefined @start_rxp
-      def to_s       
-        "#{name} := #{self.ebnf_to_s};"
-      end
-
-      def ebnf_to_s
-        if terminal?
-          "\"#{@start_rxp.source.gsub(/\\/, "")}\""
-        elsif alternation_rule?
-          terms = ebnf.term_set
-          join_rule_names(terms, " | ")
-        elsif concatenation_rule?
-          terms = ebnf.term_list
-          join_rule_names(terms, ", ")
-        elsif repetition_rule?
-          "{#{ebnf.term.rule_name}}"
-        elsif option_rule?
-          "[#{ebnf.term.rule_name}]"
-        end
-      end
-
-      def alternation_rule?
-        self.ebnf.instance_of? AlternationRule
-      end
-
-      def concatenation_rule?
-        self.ebnf.instance_of? ConcatenationRule
-      end
-
-      def repetition_rule?
-        self.ebnf.instance_of? RepetitionRule
-      end
-
-      def option_rule?
-        self.ebnf.instance_of? OptionRule
-      end
-
-      def join_rule_names(terms, infix="")
-        initial = terms.first.rule_name.to_s
-        rest = terms.drop(1)
-        rest.each_with_object(initial){|i, memo| memo.concat(infix).concat(i.rule_name.to_s)}
-      end
     end
     
     class RuleSeq < Array
